combine coarsening and spinodal decomposition
